requirements.txt
Flask-RESTful
Flask-SQLAlchemy
pytest